+++ tests/helpers/loading-from-directories/sample.type.ts
@@ -1,7 +1,0 @@
-import { Field, ObjectType } from "../../../src";
-
-@ObjectType()
-export class SampleObject {
-  @Field()
-  sampleField: string;
-}